Move RhoSpec-specific parts of TestUtil to the former
