+++ casper/src/test/scala/coop/rchain/casper/genesis/contracts/TestUtil.scala
@@ -1,50 +1,14 @@
 package coop.rchain.casper.genesis.contracts
 
-import cats.effect.{Concurrent, ContextShift, Sync}
+import cats.FlatMap
+import cats.effect.Sync
 import cats.implicits._
-import cats.{FlatMap, Parallel}
-import coop.rchain.casper.protocol.DeployData
-import coop.rchain.casper.util.ProtoUtil
 import coop.rchain.crypto.hash.Blake2b512Random
-import coop.rchain.metrics.{Metrics, Span}
 import coop.rchain.models.Par
-import coop.rchain.rholang.build.CompiledRholangSource
 import coop.rchain.rholang.interpreter.accounting.Cost
 import coop.rchain.rholang.interpreter.{NormalizerEnv, ParBuilder, Runtime}
 
 object TestUtil {
-
-  private val rhoSpecDeploy: DeployData =
-    DeployData(
-      deployer = ProtoUtil.stringToByteString(
-        "0401f5d998c9be9b1a753771920c6e968def63fe95b20c71a163a7f7311b6131ac65a49f796b5947fa9d94b0542895e7b7ebe8b91eefcbc5c7604aaf281922ccac"
-      ),
-      timestamp = 1559158671800L,
-      term = CompiledRholangSource("RhoSpecContract.rho", NormalizerEnv.Empty).code
-    )
-
-  def setupRuntime[F[_]: Concurrent: ContextShift: Metrics: Span, G[_]: Parallel[F, ?[_]]](
-      runtime: Runtime[F],
-      otherLibs: Seq[DeployData],
-      normalizerEnv: NormalizerEnv
-  ): F[Runtime[F]] =
-    for {
-      _ <- evalDeploy(rhoSpecDeploy, runtime)
-      _ <- otherLibs.toList.traverse(d => evalDeploy(d, runtime))
-      // reset the deployParams.userId before executing the test
-      // otherwise it'd execute as the deployer of last deployed contract
-      _ <- runtime.deployParametersRef.update(_.copy(userId = Par()))
-    } yield runtime
-
-  private def evalDeploy[F[_]: Sync](
-      deploy: DeployData,
-      runtime: Runtime[F]
-  ): F[Unit] = {
-    val rand: Blake2b512Random = Blake2b512Random(
-      DeployData.toByteArray(ProtoUtil.stripDeployData(deploy))
-    )
-    eval(deploy.term, runtime, NormalizerEnv(deploy))(implicitly, rand)
-  }
 
   def eval[F[_]: Sync](
       code: String,
